Rename 'helpers' module
We hereby rename the 'helpers' module to 'api_helpers'.
The idea is to collect API-related methods in such module, while using
another module for different purposes.
